refactor: 엔티티 table 제약조건 삭제
#325

+++ server/src/main/java/com/project/yozmcafe/controller/dto/menu/MenuBoardResponse.java
@@ -2,7 +2,7 @@ package com.project.yozmcafe.controller.dto.menu;
 
 import com.project.yozmcafe.domain.menu.MenuBoard;
 
-public record MenuBoardResponse(Long id, Long priority, String imageUrl) {
+public record MenuBoardResponse(Long id, int priority, String imageUrl) {
     public static MenuBoardResponse from(final MenuBoard menuBoards) {
         return new MenuBoardResponse(menuBoards.getId(), menuBoards.getPriority(), menuBoards.getImageUrl());
     }
